feat: 정산 응답 creatorName 추가  & 판매내역 강의료 Long -> BigDecimal 타입 변경 - course.java : creatorName string 필드 추가 - salerecord.amount : BigDecimal 타입으로 통일

+++ src/main/java/liveclass/creator_settlement/app/saleRecord/dto/SaleRecordCreateReq.java
@@ -4,11 +4,12 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record SaleRecordCreateReq(
     @NotBlank String courseId,
     @NotBlank String studentId,
-    @NotNull @Positive Long amount,
+    @NotNull @Positive BigDecimal amount,
     @NotNull LocalDateTime paidAt
 ) {}
